Add a cast from a uint32 to a int16 expected by eif_gen_cid
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@12836 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/traverse.c
@@ -148,7 +148,7 @@ rt_shared void traversal(char *object, int p_accounting)
 
 		/* Now insert generics */
 
-		cidarr = eif_gen_cid (dftype);
+		cidarr = eif_gen_cid ((int16) dftype);
 		i = *(cidarr++); /* count */
 
 		while (i--)
